pyproject.toml
[project]
name = "agneticrag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chromadb>=1.0.20",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "langchainhub>=0.1.21",
    "langgraph>=0.6.6",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
    "tiktoken>=0.11.0",
]
